Add GiveawaySection component
